fix: csharpsquid:S3376
